refactor classification experiments
