+++ sktime/Temp/TSF/Predictions/UnitTest/testResample0.csv
@@ -1,0 +1,25 @@
+TSF,UnitTest,test
+{'min_interval': 3, 'n_estimators': 10, 'n_jobs': 1, 'random_state': None}
+0.8181818181818182,16,0,-1,-1,2
+1,1,,0.8,0.2
+1,1,,0.9,0.1
+1,1,,0.8,0.2
+1,1,,0.8,0.2
+1,1,,0.7,0.3
+1,1,,0.9,0.1
+1,1,,0.8,0.2
+1,1,,1.0,0.0
+1,1,,1.0,0.0
+1,1,,1.0,0.0
+1,1,,0.6,0.4
+1,1,,0.9,0.1
+2,2,,0.0,1.0
+2,1,,0.5,0.5
+2,1,,0.6,0.4
+2,2,,0.1,0.9
+2,2,,0.2,0.8
+2,1,,1.0,0.0
+2,2,,0.1,0.9
+2,2,,0.1,0.9
+2,2,,0.0,1.0
+2,1,,0.6,0.4
