+++ sktime/Temp/kmeans/Predictions/UnitTest/trainResample0.csv
@@ -1,0 +1,23 @@
+kmeans,UnitTest,train
+{'averaging_algorithm': 'mean', 'averaging_algorithm_iterations': 10, 'init_algorithm': 'forgy', 'max_iter': 300, 'metric': 'dtw', 'n_clusters': 2, 'random_state': None, 'verbose': False}
+FORMAT NOT FINALISED
+1,1
+1,1
+1,1
+1,0
+1,1
+1,1
+1,0
+1,0
+1,0
+1,0
+2,1
+2,1
+2,1
+2,1
+2,0
+2,1
+2,1
+2,1
+2,1
+2,1
